// garden/frontend/src/game/guy.generated.ts
export interface GuyChild {
  depth: number;
  char: number;
  name: string | null;
  x: number;
  y: number;
  xscale: number;
  yscale: number;
  rotation: number;
  is?: string;
  mass?: number;
  minang?: number;
  maxang?: number;
  inversed?: boolean;
  p3?: string;
}